Fix marshalling of MTLResourceOptions on .NET 8

+++ src/Veldrid.MetalBindings/MTLResourceOptions.cs
@@ -1,6 +1,6 @@
 namespace Veldrid.MetalBindings
 {
-    public enum MTLResourceOptions : uint
+    public enum MTLResourceOptions : ulong
     {
         CPUCacheModeDefaultCache = MTLCPUCacheMode.DefaultCache,
         CPUCacheModeWriteCombined = MTLCPUCacheMode.WriteCombined,
